+++ R/tar_noninteractive.R
@@ -1,6 +1,6 @@
 #' @title Run if Target Markdown interactive mode is not on.
 #' @export
-#' @family utilities
+#' @family Target Markdown
 #' @description In Target Markdown, run the enclosed code
 #'   only if interactive mode is not activated. Otherwise,
 #'   do not run the code.
@@ -12,7 +12,7 @@
 #' @param code R code to run if Target Markdown interactive mode
 #'   is not turned on.
 #' @examples
-#' tar_noninteractive(list(x = 1 + 1))
+#' tar_noninteractive(message("Not in interactive mode."))
 tar_noninteractive <- function(code) {
   if_any(
     engine_knitr_is_interactive(),